+++ docs/changes.md
@@ -9,6 +9,7 @@ Auth Change Log
 * Call `Illuminate\Auth\AuthServiceProvider::boot()` during booting.
 * `Orchestra\Auth\Acl\Container` should extend `Orchestra\Memory\Abstractable\Container`.
 * Predefined package path to avoid additional overhead to guest package path.
+* Rename command to `php artisan auth:migrate`.
 
 ## Version 2.0